% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego_metrics.R
\name{calculate_ei_index}
\alias{calculate_ei_index}
\title{Calculate E-I index for ego-alter similarity based on categorical attributes}
\usage{
calculate_ei_index(
  adj_matrix,
  node_attributes,
  attribute_name,
  directed = TRUE,
  self_ties = FALSE
)
}
\arguments{
\item{adj_matrix}{An adjacency matrix representing network ties}

\item{node_attributes}{A data frame containing node attributes}

\item{attribute_name}{The name of the categorical attribute column to analyze}

\item{directed}{Logical, whether the network is directed. Default is TRUE}

\item{self_ties}{Logical, whether to include self ties. Default is FALSE}
}
\value{
A data frame with E-I index values for each ego
}
\description{
Calculate E-I index for ego-alter similarity based on categorical attributes
}
\examples{
# Create sample data
adj_mat <- matrix(c(0,1,1,0,0,0,1,0,1), nrow=3)
rownames(adj_mat) <- colnames(adj_mat) <- c("A", "B", "C")
attributes <- data.frame(
  node = c("A", "B", "C"),
  department = c("HR", "IT", "HR")
)
calculate_ei_index(adj_mat, attributes, "department")
}
