% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego_metrics.R
\name{calculate_iqv}
\alias{calculate_iqv}
\title{Calculate Agresti's Index of Qualitative Variation (IQV) for network ties}
\usage{
calculate_iqv(degree_df, node_col = "node")
}
\arguments{
\item{degree_df}{A data frame containing node degrees for different relation types}

\item{node_col}{The name of the column containing node identifiers. Default is "node"}
}
\value{
A data frame with node identifiers and corresponding IQV values
}
\description{
Calculate Agresti's Index of Qualitative Variation (IQV) for network ties
}
\examples{
# Using the output from calculate_degree
degree_data <- data.frame(
  node = c("A", "B", "C"),
  degree_friendship = c(2, 1, 3),
  degree_advice = c(1, 1, 0),
  degree_conflict = c(0, 2, 1)
)
calculate_iqv(degree_data)
}
