% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego_metrics.R
\name{summarize_valued_ties}
\alias{summarize_valued_ties}
\title{Summarize valued tie data for each node in a network}
\usage{
summarize_valued_ties(graph, weight_attr = "weight", mode = "all")
}
\arguments{
\item{graph}{An igraph object with weighted edges}

\item{weight_attr}{The name of the edge attribute containing tie weights. Default is "weight"}

\item{mode}{Character, the type of edges to consider: "in", "out", or "all". Default is "all"}
}
\value{
A data frame with summary statistics for each node's tie strengths
}
\description{
Summarize valued tie data for each node in a network
}
\examples{
# Create a weighted network
library(igraph)
g <- graph_from_edgelist(matrix(c(1,2,1,3,2,3,2,4,3,4), ncol=2, byrow=TRUE), directed=FALSE)
E(g)$weight <- c(3, 1, 2, 4, 5)
V(g)$name <- paste0("A", 1:4)
summarize_valued_ties(g)
}
